# implement ptr::write without dedicated intrinsic