… src/test/codegen/intrinsics/move-val-init.rs
… src/test/ui/intrinsics/intrinsic-move-val-cleanups.rs
… src/test/ui/intrinsics/intrinsic-move-val.rs
… src/test/ui/unsafe/unsafe-move-val-init.rs
… src/test/ui/unsafe/unsafe-move-val-init.stderr