@RalfJung 2020-12-22
Let's see if perf says anything.

@bors try