@bjorn3 2020-12-25
Perf results for https://github.com/ebobby/simple-raytracer/commit/804a7a21b9e673a482797aa289a18ed480e4d813:

```
Benchmark #1: ./raytracer_75e1acb
  Time (mean ± σ):      8.040 s ±  0.019 s    [User: 8.037 s, System: 0.004 s]
  Range (min … max):    8.013 s …  8.065 s    10 runs
 
Benchmark #2: ./raytracer_63c71d4
  Time (mean ± σ):      8.086 s ±  0.022 s    [User: 8.083 s, System: 0.003 s]
  Range (min … max):    8.061 s …  8.138 s    10 runs
 
Summary
  './raytracer_75e1acb' ran
    1.01 ± 0.00 times faster than './raytracer_63c71d4'
```

The perf difference is so small as simple-raytracer only spends a tiny bit of time in `std::ptr::write`. The codegen is noticably worse though.

Before:

```
       │    000000000007ab40 <core::ptr::write>:
       │    _ZN4core3ptr5write17h8616c78d24e53297E():
 32,01 │      sub    $0x18,%rsp
 19,97 │      mov    %rdi,0x8(%rsp)
 16,00 │      mov    %rsi,0x10(%rsp)
 32,02 │      mov    %rsi,(%rdi)
       │      add    $0x18,%rsp
       │    ← retq
```

After:

```
       │    000000000007ac00 <core::ptr::write>:
       │    _ZN4core3ptr5write17h8616c78d24e53297E():
 18,93 │      sub    $0x28,%rsp
  8,13 │      mov    %rsi,(%rsp)
 18,96 │      mov    %rdi,0x10(%rsp)
       │      movb   $0x0,0xf(%rsp)
  5,42 │      movb   $0x1,0xf(%rsp)
 10,72 │      mov    (%rsp),%rax
 29,71 │      mov    %rax,(%rdi)
  8,13 │      movb   $0x0,0xf(%rsp)
       │      add    $0x28,%rsp
       │    ← retq 
```